+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 -------------
 Version (dev)
 -------------
+* Add :mod:`srctools.sndscape`, for parsing soundscape scripts.
 * Changed `PackList.load_choreo_scenes() <srctools.packlist.PackList.load_choreo_scenes>` to warn if ``scenes.image`` is not found.
 
 -------------